+++ atmega32u4/hardware.c
@@ -1,6 +1,10 @@
 // Copyright 2018 jem@seethis.link
 // Licensed under the MIT license (http://opensource.org/licenses/MIT)
 
+#include "core/hardware.h"
+
+#include <avr/wdt.h>
+
 void led_init(void) {
 }
 
@@ -8,16 +12,33 @@ void reset_mcu(void) {
 
 }
 
-void bootloader_jmp(void) {
+#if defined(BOOTLOADER_ATMEL_DFU)
 
+#include "bootloaders/atmel_dfu/atmel_bootloader.h"
+
+void bootloader_jmp(void) {
+    run_bootloader();
 }
 
-void bootloader_jmp_2(void) {
+#elif defined(BOOTLOADER_KP_BOOT_32U4)
 
+#include "bootloaders/kp_boot_32u4/interface/kp_boot_32u4.h"
+
+void bootloader_jmp(void) {
+    cli();
+    wdt_enable(WDTO_15MS);
+    wdt_reset();
+    *(uint16_t*)MAGIC_ADDRESS = MAGIC_ENTER_BOOT;
+    while (1);
+}
+#endif
+
+void wdt_init(void) {
+    wdt_enable(WDTO_1S);
 }
 
 void wdt_kick(void) {
-
+    wdt_reset();
 }
 
 void led_testing_set(void) {