+++ src/core/bootloader.h
@@ -4,5 +4,4 @@
 #pragma once
 
 void bootloader_jmp(void);
-
 void bootloader_jmp_2(void);