+++ CHANGELOG.md
@@ -27,6 +27,8 @@ Internal changes:
 * `keyplus-cli` add commands to dump the raw hex of the layout and settings section
 * `keyplus-cli` is now able to update RF, layout, and ID separately.
 
+* `keyplus-flasher` add support for kp_boot_32u4 bootloader
+
 * `firmware` add new `chip_id` concept to firmware settings allowing the
   `python-api` to discover what microcontroller the keyboard is using.
 * `firmware` rework flash writing commands so that is possible to update only a